% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_Proximal.R
\docType{package}
\name{AlgoProximal-package}
\alias{AlgoProximal}
\alias{AlgoProximal-package}
\title{AlgoProximal: Proximal algorithms for Lasso}
\description{
Implementations and pedagogical examples of ISTA for the Lasso problem.
(Les détails principaux sont dans DESCRIPTION et dans la vignette.)
}
\author{
\strong{Maintainer}: TCHATCHOUANG DJUICHI AELLE \email{aelle1672@gmail.com}

}
\keyword{internal}
