% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipschitz_constant.R
\name{calculate_lipschitz_const}
\alias{calculate_lipschitz_const}
\title{Lipschitz Constant For Least Squares}
\usage{
calculate_lipschitz_const(X)
}
\arguments{
\item{X}{predicator matrice}
}
\value{
The lipschitz constant
}
\description{
Calculate L =max(eigen(X^T X /N)),to define the time step 1/L
}
