% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_loss.R
\name{grad_loss}
\alias{grad_loss}
\title{The Loss Gradient function for least squares}
\usage{
grad_loss(X, y, beta)
}
\arguments{
\item{X}{Predicator matrice}

\item{y}{Vector of responses}

\item{beta}{Vector current coefficient}
}
\value{
A numerical vector ,gradient of f(beta)
}
\description{
Calculate the gradient of the smooth term for ISTA iteration
}
