% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISTA_LASSO.R
\name{ista_lasso}
\alias{ista_lasso}
\title{Solver LASSO by ISTA Proximal Algorithm}
\usage{
ista_lasso(X, y, lambda, max_iter = 1000, tol = 1e-04)
}
\arguments{
\item{X}{Predicator matrice}

\item{y}{Vector of responses}

\item{lambda}{The LASSO regularisation parameter}

\item{max_iter}{Number of iterations}

\item{tol}{Convergence tolerance}
}
\value{
A list which contains the final coeffiients and the number of iterations
}
\description{
Implement the ISTA loop to minimise the LASSO objectif
}
