% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_l1.R
\name{prox_l1}
\alias{prox_l1}
\title{Proximal Operator For l1 Norm(Soft-Thresholding)}
\usage{
prox_l1(x, tau)
}
\arguments{
\item{x}{A numerical vector on which we apply the operator.}

\item{tau}{The effectif threshold(lambda * step_size)}
}
\value{
A resultant numerical vector
}
\description{
Implement the operation sign(x) * pmax(0,abs(x)-tau)
}
